% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ws_close}
\alias{ws_close}
\title{ws_close}
\usage{
ws_close(ws_ptr)
}
\arguments{
\item{Chrome}{ws ptr object.
Using this function leads to broken sockets on MacOS. Do not use.}
}
\description{
Close a socket connection
}

