% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ws_poll}
\alias{ws_poll}
\title{Consume one message}
\usage{
ws_poll(ws_ptr, timeout = 5L)
}
\arguments{
\item{ws_ptr}{pointer from \code{\link{ws_connect}}}

\item{timeout}{time before timing out each time server is queried}
}
\value{
string with message from server
}
\description{
Consume one message
}
\examples{
\dontrun{
    con <- ws_connect("ws://localhost:5006/") # Need to have websocket server
    ws_poll(con, 5) # returns one message
}

}
