% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ws_poll_list}
\alias{ws_poll_list}
\title{Consume n events}
\usage{
ws_poll_list(ws_ptr, eventlimit)
}
\arguments{
\item{eventlimit}{limit of events}
}
\value{
list of strings with each message
}
\description{
Consume n events
}
