% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ws_read_one}
\alias{ws_read_one}
\title{ws_read_one}
\usage{
ws_read_one(ws_ptr, timeout = 5L)
}
\arguments{
\item{ws_ptr}{pointer from \code{\link{ws_connect}}}

\item{timeout}{time in milliseconds before timing out, see details}
}
\value{
one message, if one occurs within timeout. otherwise, empty string
}
\description{
Consume one message within timeout
}
\details{
If timout is set to a negative number, the process should
block until a message arrives. If set to 0, no blocking is performed --
a message is returned if it is there.  If set to a positive value,
the function will wait until there is a message or a timeout
}
\examples{

\dontrun{
    con <- ws_connect("ws://localhost:5006/") # Need to have websocket server
    ws_read_one(con, 5)
}
}
