% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ws_receive_multiple}
\alias{ws_receive_multiple}
\title{ws_receive_multiple}
\usage{
ws_receive_multiple(ws_ptr, eventlimit)
}
\arguments{
\item{ws_ptr}{pointer from \code{\link{ws_connect}}}

\item{eventlimit}{limit of events}
}
\value{
list of strings with each message
}
\description{
Consume n events
}
\examples{

\dontrun{
    con <- ws_connect("ws://localhost:5006/") # Need to have websocket server
    ws_receive_multiple(con, 3) # returns 3 messages
}
}
