% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ws_receive_one}
\alias{ws_receive_one}
\title{ws_receive_one}
\usage{
ws_receive_one(ws_ptr, frequency = 5L)
}
\arguments{
\item{ws_ptr}{pointer from \code{\link{ws_connect}}}

\item{frequency}{time before timing out each time server is queried}
}
\value{
string with message from server
}
\description{
Query websocket until one message received
}
\examples{
\dontrun{
    con <- ws_connect("ws://localhost:5006/") # Need to have websocket server
    ws_receive_one(con, 5) # returns one message
}

}
